+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/response/order/QueryOrderProductListRespVO.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -40,13 +41,13 @@ public class QueryOrderProductListRespVO {
     private String unitName;
 
     @ApiModelProperty("单价")
-    private Long price;
+    private BigDecimal price;
 
     @ApiModelProperty("数量")
     private Long num;
 
     @ApiModelProperty("总价")
-    private Long amount;
+    private BigDecimal amount;
 
     @ApiModelProperty("订单编码(订单号)")
     private String orderCode;
